O(1) solution for the four main methods base on doubly-linked-list
