+++ 01-not-an-array.js
@@ -26,19 +26,97 @@
 //  notAnArray.shift();
 //  returns 'now I am first' and the object should look like {0: 'first'}
 //
+class LinkedListNode {
+  constructor(v, prev, next) {
+      this.v = v;
+      this.prev = prev;
+      this.next = next;
+  }
+}
 
 class NotAnArray {
   constructor() {
-    this.storage = {};
+    this.head = null;
+    this.tail = null;
   }
 
-  push() {}
+  // add node to end of the list
+  push(v) {
+    const node = new LinkedListNode(v);
 
-  pop() {}
+    if (this.isEmpty()) {
+      this.head = node;
+      this.tail = node;
+    } else {
+      node.prev = this.tail;
+      this.tail.next = node;
+      this.tail = node;
+    }
 
-  unshift() {}
+    return this;
+  }
 
-  shift() {}
+  pop() {
+    let last = this.tail;
+
+    if (last) {
+      this.tail = last.prev;
+
+      if (this.tail) {
+        this.tail.next = null;
+      }
+
+      if (this.head === last) {
+        this.head = null;
+      }
+    }
+
+    return (last) ? last.v : undefined;
+  }
+
+  unshift(v) {
+    let node = new LinkedListNode(v);
+
+    node.next = this.head;
+    this.head = node;
+
+    return this;
+  }
+
+  shift() {
+    let node;
+
+    if (this.head) {
+      node = this.head;
+      this.head = this.head.next;
+      if (node === this.tail) {
+        // single item
+        this.tail = this.head;
+      }
+    }
+
+    return (node) ? node.v : undefined;
+  }
+
+  // auxillaries
+  isEmpty() {
+      return this.head == null; // treat undefined as null
+  }
+
+  get storage() {
+    let p = this.head;
+    let vals = {};
+    let i = 0;
+    while(p) {
+        vals[i] = p.v;
+        p = p.next;
+        i += 1;
+    }
+
+    return vals;
+  }
+
+  set storage(o) {}
 }
 
 module.exports = NotAnArray;
